Remove useless conversion to/from world units
The `syncRequest` / `eventSyncRequest` is exclusive to JS API, so removed the useless conversion to world units and subsequent conversion back to tile units. Entire process is now in tile units. Also, as maps are limited to 256x256 tiles, changed `int32` to `uint8` to reduce packet size.

+++ src/multiplay.h
@@ -401,7 +401,7 @@ bool shouldSkipReadyResetOnPlayerJoinLeaveEvent();
 STRUCTURE *findResearchingFacilityByResearchIndex(const PerPlayerStructureLists& pList, unsigned player, unsigned index);
 STRUCTURE *findResearchingFacilityByResearchIndex(unsigned player, unsigned index); // checks apsStructLists
 
-void sendSyncRequest(int32_t req_id, int32_t x, int32_t y, const BASE_OBJECT *psObj, const BASE_OBJECT *psObj2);
+void sendSyncRequest(int32_t req_id, uint8_t x, uint8_t y, const BASE_OBJECT *psObj, const BASE_OBJECT *psObj2);
 
 
 bool sendBeaconToPlayer(SDWORD locX, SDWORD locY, SDWORD forPlayer, SDWORD sender, const char *beaconMsg);
